added check for ip
