+++ src/components/navbar/DiscoveryForm.jsx
@@ -14,11 +14,36 @@ const DiscoveryForm = ({ handleSubmit, onCancel }) => {
         }));
     };
 
+    const isValidIPv4 = (ip) => {
+        const ipv4Pattern =
+            /^(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)$/;
+        return ipv4Pattern.test(ip);
+    };
+
+    const isValidCIDR = (cidr) => {
+        const cidrPattern =
+            /^(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\/([0-9]|[1-2][0-9]|3[0-2])$/;
+        return cidrPattern.test(cidr);
+    };
+
+    const areAllIPAddressesValid = (input) => {
+        const ipAddresses = input.split(",").map((ip) => ip.trim());
+        return ipAddresses.every((ip) => isValidIPv4(ip) || isValidCIDR(ip));
+    };
+
     const handleFormData = () => {
         if (formData.address.trim().length === 0) {
             formData.discover_from_config = true;
         }
-        handleSubmit(formData);
+
+        if (areAllIPAddressesValid(formData.address)) {
+            // handleSubmit(formData);
+            console.log(formData);
+        } else {
+            alert("Invalid IP Address");
+            return;
+        }
+
     };
 
     return (
@@ -38,12 +63,20 @@ const DiscoveryForm = ({ handleSubmit, onCancel }) => {
                     onChange={handleChange}
                     placeholder="Network Address or Device IP."
                 />
+
+                <p className="mt-10">
+                    Note : Use (,) Comma to separate the multiple IP address{" "}
+                </p>
             </div>
             <div className="">
                 <button className="btnStyle mt-10 mr-10" type="submit">
                     Submit
                 </button>
-                <button className="btnStyle mt-10" type="button" onClick={onCancel}>
+                <button
+                    className="btnStyle mt-10"
+                    type="button"
+                    onClick={onCancel}
+                >
                     Cancel
                 </button>
             </div>
